requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
firebird-driver==1.10.10
python-dotenv==1.0.0
email-validator==2.0.0
Pillow==10.0.1
python-dateutil==2.8.2
